chore(user-interviews): revert contracts move, restore canonical location
CI's "Lint product structure" check (tools/hogli-commands/.../product/checks.py)
requires every product listed in tach.toml's canonical facade alternation to
have backend/facade/contracts.py — moving contracts.py to backend/contracts.py
broke that invariant.

This matches products/architecture.md, which puts contracts.py under facade/,
and the established pattern in mcp_analytics, experiments, and access_control
where logic.py also imports from facade/contracts.py.

The greptile bot's layering concern is noted but doesn't survive the codebase
convention: facade/contracts.py is a public-surface marker (no Django imports,
tracked as Turbo contract inputs), not the outer wrapping layer it might look
like from the path alone.

Generated-By: PostHog Code
Task-Id: 639c774a-3c16-4b58-94a4-35682c62ade2
